refactor: HttpResponseBuilder 생성
